Fix for a total page issue
We took all the "a" tag contained in the page div at the bottom. There
was an incrementation of the final number of page because the current
one was not a "a" tag. Nevertheless, now, there is new "a" tag
corresponding to the arrow to go to the next page. So we don't need to
add 1 to the number of page anymore.

The issue was that for the last page which doesn't exists, we get links
of random contents...